pyproject.toml
[tool.poetry]
name = "text-to-image"
version = "0.1.0"
description = ""
authors = ["Naman Malik <namanmalik06@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15.2"
pillow = "^10.4.0"
gunicorn = "^22.0.0"
django-cors-headers = "^4.4.0"
psycopg2-binary = "^2.9.9"
celery = "^5.4.0"
redis = "^5.0.8"
black = "^24.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
